refactor: move models.py to models folder and change name to course_data_model.py
